+++ src/OpenTelemetry.Exporter.Prometheus.HttpListener/PrometheusHttpListenerMeterProviderBuilderExtensions.cs
@@ -71,6 +71,7 @@ public static class PrometheusHttpListenerMeterProviderBuilderExtensions
             ScrapeResponseCacheDurationMilliseconds = options.ScrapeResponseCacheDurationMilliseconds,
             TargetInfoEnabled = options.TargetInfoEnabled,
             DisableTotalNameSuffixForCounters = options.DisableTotalNameSuffixForCounters,
+            MaxScrapeResponseSizeBytes = options.MaxScrapeResponseSizeBytes,
         });
 
         var reader = new BaseExportingMetricReader(exporter)